Fixing urgent security vulnerabilities
Fixed all security vulnerabilities apart from Lodash which is a Grunt
transitive dependency. We have the latest version of Grunt installed so
there's not a lot we can do. The only only consolation is that Grunt
is a dev dependency, therefore, it wouldn't be in production.

The only real solution would be to pull out Grunt (our build tool) and
replace that with Webpack. Grunt is an antiquated build tool whereas
Webpack is current and the preferred approach.

Most dependencies have been upgraded apart from the following:

Package                Current      Wanted         Latest
axe-core                 2.6.1       2.6.1          3.3.1
axe-webdriverjs          1.3.0       1.3.0          2.3.0
winston                  2.4.1       2.4.1          3.2.1
json-schema-faker   0.5.0-rc10  0.5.0-rc10     0.5.0-rc17
selenium-webdriver       3.6.0       3.6.0  4.0.0-alpha.4

The top 2 dependencies couldn't be upgraded as the integration tests
were failing on CircleCI - there are no security vulnerabilities with
the current version.

The Winston dependency has considerable breaking changes when upgrading
to 3.2.1 which would take some time to fix, we should revisit this at a
suitable time - there are no security vulnerabilities with the current
version.

The last two dependencies are either a release candidate or in alpha
which we don't care about.